# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ECHO Anonymous Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserByUsername
      tags: [users]
      summary: Search user by username and get public key
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUser"
        "404":
          description: User not found
...
  /users/me/settings:
    patch:
      operationId: updateSettings
      tags: [users]
      summary: Update user settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /prekeys:
    post:
      operationId: uploadPrekeys
      tags: [prekeys]
      summary: Upload new one-time pre-keys
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrekeysInput"
      responses:
        "200":
          description: Pre-keys uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /messages:
    get:
      operationId: getMessages
      tags: [messages]
      summary: Get messages for a chat
      parameters:
        - name: chatId
          in: query
          required: true
          schema:
            type: integer
        - name: before
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send an encrypted message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /messages/{id}/read:
    post:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /files/upload:
    post:
      operationId: uploadFile
      tags: [files]
      summary: Upload encrypted file (base64 encoded)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileUploadInput"
      responses:
        "200":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUploadResult"
  /files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Download file metadata
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileInfo"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    PublicUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        publicIdentityKey:
          type: string
        publicSignedPrekey:
          type: string
        oneTimePrekey:
          type: ["string", "null"]
        lastActive:
          type: ["string", "null"]
        avatarFileId:
          type: ["string", "null"]
      required: [id, username, publicIdentityKey]
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        publicIdentityKey:
          type: string
        lastActive:
          type: string
        avatarFileId:
          type: ["string", "null"]
        settings:
          $ref: "#/components/schemas/UserSettings"
        createdAt:
          type: string
      required: [id, username, publicIdentityKey, lastActive, createdAt]
    UserSettings:
      type: object
      properties:
        theme:
          type: string
        language:
          type: string
        notifications:
          type: boolean
        showLastSeen:
          type: boolean
        showAvatar:
          type: boolean
        autoDeleteMessages:
          type: ["integer", "null"]
    SettingsUpdate:
      type: object
      properties:
        settings:
          $ref: "#/components/schemas/UserSettings"
    PrekeysInput:
      type: object
      properties:
        publicOneTimePrekeys:
          type: array
          items:
            type: string
      required: [publicOneTimePrekeys]
    Message:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        senderUsername:
          type: string
        chatId:
          type: integer
        chatType:
          type: integer
        encryptedContent:
          type: string
        timestamp:
          type: string
        deliveredAt:
          type: ["string", "null"]
        readAt:
          type: ["string", "null"]
        deleteAfterRead:
          type: boolean
        deleteAfterSeconds:
          type: ["integer", "null"]
        isVoice:
          type: boolean
        mediaFileId:
          type: ["string", "null"]
        replyToId:
          type: ["integer", "null"]
      required: [id, senderId, chatId, chatType, encryptedContent, timestamp, deleteAfterRead, isVoice]
    MessageInput:
      type: object
      properties:
        receiverId:
          type: ["integer", "null"]
        chatId:
          type: ["integer", "null"]
        chatType:
          type: integer
        encryptedContent:
          type: string
        deleteAfterRead:
          type: boolean
        deleteAfterSeconds:
          type: ["integer", "null"]
        isVoice:
          type: boolean
        mediaFileId:
          type: ["string", "null"]
        replyToId:
          type: ["integer", "null"]
      required: [chatType, encryptedContent]
...
    FileUploadInput:
      type: object
      properties:
        data:
          type: string
          description: Base64-encoded encrypted file data
        mimeType:
          type: string
        encryptedKey:
          type: string
      required: [data]
    FileUploadResult:
      type: object
      properties:
        fileId:
          type: string
        url:
          type: string
        size:
          type: integer
        mimeType:
          type: string
      required: [fileId, url, size]
    FileInfo:
      type: object
      properties:
        fileId:
          type: string
        size:
          type: integer
        mimeType:
          type: string
        createdAt:
          type: string
      required: [fileId, size, createdAt]
...