# lib/api-zod/src/generated/types/fileInfo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface FileInfo {
  fileId: string;
  size: number;
  mimeType?: string;
  createdAt: string;
}