# lib/api-zod/src/generated/types/fileUploadInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface FileUploadInput {
  /** Base64-encoded encrypted file data */
  data: string;
  mimeType?: string;
  encryptedKey?: string;
}