# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  /** @nullable */
  receiverId?: number | null;
  /** @nullable */
  chatId?: number | null;
  chatType: number;
  encryptedContent: string;
  deleteAfterRead?: boolean;
  /** @nullable */
  deleteAfterSeconds?: number | null;
  isVoice?: boolean;
  /** @nullable */
  mediaFileId?: string | null;
  /** @nullable */
  replyToId?: number | null;
}