# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface PublicUser {
  id: number;
  username: string;
  publicIdentityKey: string;
  publicSignedPrekey?: string;
  /** @nullable */
  oneTimePrekey?: string | null;
  /** @nullable */
  lastActive?: string | null;
  /** @nullable */
  avatarFileId?: string | null;
}