# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */
import type { UserSettings } from './userSettings';

export interface SettingsUpdate {
  settings?: UserSettings;
}