# lib/api-zod/src/generated/types/userSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ECHO Anonymous Messenger API
 * OpenAPI spec version: 0.1.0
 */

export interface UserSettings {
  theme?: string;
  language?: string;
  notifications?: boolean;
  showLastSeen?: boolean;
  showAvatar?: boolean;
  /** @nullable */
  autoDeleteMessages?: number | null;
}